Update Kernel to be a decimal values. Default background come to transparent

+++ DigitalDoor.Reporting.Entities/ValueObjects/Kernel.cs
@@ -5,15 +5,15 @@
     /// </summary>
     public class Kernel
     {
-        public int Top { get { return TopBK; } set { TopBK = value; } }
-        private int TopBK;
-        public int Right { get { return RightBK; } set { RightBK = value; } }
-        private int RightBK;
+        public decimal Top { get { return TopBK; } set { TopBK = value; } }
+        private decimal TopBK;
+        public decimal Right { get { return RightBK; } set { RightBK = value; } }
+        private decimal RightBK;
 
-        public int Bottom { get { return BottomBK; } set { BottomBK = value; } }
-        private int BottomBK;
-        public int Left { get { return LeftBK; } set { LeftBK = value; } }
-        private int LeftBK;
+        public decimal Bottom { get { return BottomBK; } set { BottomBK = value; } }
+        private decimal BottomBK;
+        public decimal Left { get { return LeftBK; } set { LeftBK = value; } }
+        private decimal LeftBK;
 
         public Kernel()
         {
@@ -23,9 +23,9 @@
             LeftBK = 0;
 
         }
-        public Kernel(int allPoint) : this(allPoint, allPoint, allPoint, allPoint) { }
-        public Kernel(int top, int left) : this(top, 0, 0, left) { }
-        public Kernel(int top, int right, int bottom, int left) : this() =>
+        public Kernel(decimal allPoint) : this(allPoint, allPoint, allPoint, allPoint) { }
+        public Kernel(decimal top, decimal left) : this(top, 0, 0, left) { }
+        public Kernel(decimal top, decimal right, decimal bottom, decimal left) : this() =>
             (TopBK, RightBK, BottomBK, LeftBK) = (top, right, bottom, left);
